+++ src/atoms/buttons/Button/Button.stories.ts
@@ -4,7 +4,7 @@ import Button from "./Button";
 
 // More on how to set up stories at: https://storybook.js.org/docs/react/writing-stories/introduction#default-export
 const meta = {
-  title: "Atoms/Button",
+  title: "Atoms/Buttons/Button",
   component: Button,
   parameters: {
     // Optional parameter to center the component in the Canvas. More info: https://storybook.js.org/docs/react/configure/story-layout
